# api/api.yml
...
openapi: "3.0.3"
info:
  version: 1.0.0
  title: Argos Service
  license:
    name: GPL v3
servers:
  - url: /api
security:
  - bearerAuth: []
paths:
...
  /nodes/{parentId}/supplychains:
    post:
      summary: create a supplychain
      operationId: createSupplyChain
      tags:
        - supply chain
      parameters:
        - name: parentId
          in: path
          required: true
          description: this will be the project id
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplyChain"
      responses:
        '201':
          description: created
          headers:
            Location:
              schema:
                type: string
                format: uri
              description: location where the created supplychain can be found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplyChain"
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/Error'
    get:
      summary: get supplychains for a parent
      operationId: getSupplyChains
      tags:
        - supply chain
      parameters:
        - name: parentId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        '200':
          description: the supply chain
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupplyChain"
        '404':
          description: not found
        '500':
          $ref: '#/components/responses/Error'
  /supplychains/{supplyChainId}:
    get:
      summary: get a supplychain
      operationId: getSupplyChain
      tags:
        - supply chain
      parameters:
        - name: supplyChainId
          in: path
          description: supply chain id
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        '200':
          description: the supply chain item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplyChain"
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/Error'
    put:
      summary: update supplychain by id
      operationId: updateSupplyChain
      tags:
        - supply chain
      parameters:
        - name: supplyChainId
          in: path
          description: supply chain id
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplyChain"
      responses:
        '200':
          description: updated supply chain
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplyChain'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/Error'
    delete:
      summary: delete supply chain by id. The layout, approval configs and link objects are also deleted
      operationId: deleteSupplyChainById
      tags:
        - supply chain
      parameters:
        - name: supplyChainId
          in: path
          description: supply chain id
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        '204':
          description: supply chain deleted
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
  responses:
    ValidationError:
      description: validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
    Error:
      description: server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    UUID:
      type: string
      format: uuid
...
    NameString:
      type: string
      minLength: 1
      maxLength: 63
      pattern: '^([a-z]|[a-z][a-z0-9-]*[a-z0-9])?$'
...
    SupplyChain:
      type: object
      required:
        - name
        - parentId
      properties:
        id:
          $ref: "#/components/schemas/UUID"
          readOnly: true
        name:
          $ref: "#/components/schemas/NameString"
        pathToRoot:
          type: array
          maxItems: 256
          items:
            $ref: "#/components/schemas/UUID"
        parentId:
          $ref: "#/components/schemas/UUID"          
...
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    ValidationError:
      type: object
      required:
        - messages
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ValidationMessage"
    ValidationMessage:
      type: object
      required:
        - message
        - type
      properties:
        field:
          type: string
        type:
          enum:
            - DATA_INPUT
            - MODEL_CONSISTENCY
        message:
          type: string
...